gnubik: spell out the colours and the six faces
0 red       Front  F
1 green     Back      B
2 blue      Up     U
3 cyan      Down      D
4 magenta   Left   L
5 yellow    Right     R

Notes: there are 3x3x3 cubes, with
 8  corner cubes that are part of 3 faces
 12 edge cubes that are part of 2 faces
 6  centre cubes that are part of 1 face, and
 1  invisible cube at the centre of the Rubic,
    with no face.

Things, here, are abstraced within the concepts of a
non-abelian group.  The set of 12 basic elements is
 {F,f,B,b,U,u,D,d,L,l,R,r}.

 F clockwise 90 degree rotation of the front face
 f counter-clockwise
 B clockwise back face rotation by 90 degrees
 b count-clockwise, back face
 U up face, clockwise
 u up face, counter-clockwise
 D down face, clock
 d down face, counter-clock
 L left face, clock
 l left face, counter clock
 R right face, clock
 r right face, counter-clock

Names of the 26 visible cubes' examples
 F   centre cube that is visible
     on the front face (red)
 FU  edge cube seen
     on the front and up faces (red,blue)
 FUL corner cube seen on
     front, up, and left faces (red,blue,magenta)

The Rubik group's zero element: no rotation, and
no change to the Rubik's arrangement, 0 (for now).

The group's inverse element of the F transform, for
example, being the rotation of the front face in the
clockwise direction, is the f transform,
counter-clockwise rotation of the front face.

The group's additive operation * sequences
two Rubik elements, and creates another composite
element of the Rubik group's set.

For example, connecting two basic elements F and U,
creates the composite Rubik element F*U. This
operation first rotates the front face by
90 degrees in the clockwise direction, and then
rotates of the up face by 90 degrees. (Note:
order of operations matter as the group is
non-commutative.)

As another example, the composite element A=F*U
of the Rubik group has the inverse element -A=u*f.
Proof: A*-A=0, with 0 being the neutral element,
since A*-A=F*U*(u*f)=0, and no change of
theRubik's arrangement is effected.

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>