Rubik: a: review
(what is here, so far?)

Fn/read-logs-of-interest -- a shell function,
                            git log messages,
			    on the subjecct of the Rubik group

The Rubik cube itself, the Rubik group to abstract (R,*,0,-)
the set R, the binary operation *, the neutral element that
does nothing, the inverse of A being -A.

Permutations, canonical cycle notation, two line notation,
appying the group operation * to group elements to make new
elements, algorithm to write down a permutation (an element)
in canonical cycle notation.

Updating this git log, the shell function Fn/git-commit-amend-here.

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>