Rubik: a: counting woozy beans
==============================================
% 20220723 (C) Gunter Liszewski -*- mode: tex; -*-
% permutation basics

\def\a{\left(\matrix}
\def\b{\cr}
\def\c{\right)}

\def\d{\footnote*{The Mathematics of the Rubik’s Cube,
  http://web.mit.edu/sp.268/www/rubik.pdf}}
\def\f{\footnote*{TAoCP chapter 1.2.5,
  https://www-cs-faculty.stanford.edu/\%7Eknuth/taocp.html}}
\def\g{\footnote*{\TeX book, texbook.tex,
  https://www.ctan.org/tex-archive/systems/knuth/dist/tex}}
%
\def\e{\bigskip}
\def\R{Woozy}
\def\*{\circ}
%
% 595.28 841.89 a5: 419.53 595.28
\vsize=300pt % trial and likely error
\hsize=480pt % landscape to view on your screen
\footline={\tenrm Permutations\quad\dotfill\quad \folio}
\raggedright

\beginsection 1. Plain \TeX nology % and some Maths

\proclaim Theorem T. All things are not necessarily the same\g

\beginsection 2. Permutations

\e
TAoCP in chapter 1.2.5 gives two methods to generate
all permutations of a given ordered set.
Quantites of permutations are considered
with relevance to computing efficiencies.

\beginsection 3. The Wide-Awake example Group

\e
We re-think\d, re-word, and re-start
with a set of attributes, elements or objects,\break
$W=\{$ woozy, vacuous, sleepy, wide-awake $\}.$
These elements are used to generate all possible
arrangements $\eta$ which are orderd $n$-tuples with
$1\le n\le4.$
For example, $\eta=($ woozy, wide-awake $)$ is a $2$-tuple.
Now the set $\R$ is the set of all permutations that jumble
such elements like $\eta.$

\e
Let $(\R,\*,0,-)$ be the group with the set $\R$, a binary
operation $\*$, a neutral elment $0$, and
for each element $\pi\in\R$ there is an inverse element
$-\pi\in\R$ such that $\pi\*-\pi=0.$

\e
For now, here, we call this group's binary operation $composition$.
Given two elements
$\pi,\eta\in\R,$ then
$\pi\*\eta\in\R$ and $\eta\*\pi\in\R.$

% \e
% in a woozy state of progress
% $\pi=(a_1 a_2 ... a_n),$ and $\eta=(b_1 b_2 ... b_n)$, then
% $\pi\*\eta\in\R$ and $\eta\*\pi\in\R.$

\eject

\beginsection 4. Creating the Woozy set

\proclaim Theorem X. An ordered set of $n$ elements
has $n!$~arrangements.

This had a little consideration.  Here, we
convey our understanding of the Permutations and
Factorials section.\f

Given a set of objects $W=\{a_1,a_2,...,a_n\}$.
$P_n$ is the set of arrangements given $n$ objects
$a_1,...,a_n$ $\in W$, such as
$\{(a_1,a_2,...a_n),(a_2,a_1,...),...\}$. For example,
with $W=\{1,2,3\}$, we have
$$P_3=\{(123),(231),(312),(132),(321),(213)\}.$$

\e
Method 1, now, moves from $n=3$ to $n=4$ as follows.
For each element in $P_{n-1}=P_3$, place element
$a_n$ in each possible vacuous position to arrive at
$P_n=P_4$, that is
$$P_4=\{(a_na_1a_2a_3),
       (a_1a_na_2a_3),
       (a_1a_2a_na_3),
       (a_1a_2a_3a_n),
       ...,
       (a_na_2a_1a_3),
       (a_2a_na_1a_3),
       (a_2a_1a_na_3),
       (a_2a_1a_3a_n)\}$$

\beginsection 5. Accounting for these Arrangements

\e
Adding up all permutations that are so generated we have
$p_n$ the number of all elements in $P_n$ % to be
%$$p_n=\sum_{k=1}^np_k.$$.

\e
And again, after some re-view, we sense a need to re-word.
$P_{nn}$ is the set of permuted $n$-tuples, and $P_n$ is
the, probably bigger, set of all the $k$-tuples with
$k\in\{1,2,..,n\}.$. In other words, $P_n$
may mean different things, or sets of things. This also
applies to quantities that could be denoted like $p_{nk},$
and $p_{nn},$ and in case of our big wide-awake bean bag,
which we sum up to $p_n;$  probably.

\e
First, we started with $p_n=\sum_{k=1}^n k!$
to be the quantity
$p_n$ that accounts for all the elements of
arrangements in set $P_n$,
with $p_k=k!$ for $1\le k\le n$.

However, on the back of some scrap paper, we jotted down
$\{(1),(2),(3),(4)\}$ and saw that $\{(2),(3),(4)\}$ are
not included in our sum, and
$\{(12),(21),(13),(31),(14),(41),$
$(23),(32),(24),(42),(34),(43)\}$ has $10$ 2-tuples unaccounted
for, etc.)

\e
So, for now, given that $p_{nk}=n(n-1)...(n-k+1)$\f,
combined with $p_n=\sum_{k=1}^n p_{nk}$,
we count the number of arrangements of $n$ objects to be
$p_n=\sum_{k=1}^n {n!\over{(n-k)!}}$
or some such like.
\bye
==============================================
(under consideration)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>