admin: git: rotating two faces, F and R
permutating of F, R or both
==================================================
\beginsection 10. The example 2F 2R cycle and its abstraction

\def\R{\hbox{\bf R}}%
\e Spinning the cube.  Let $\pi\in S$ and $S\subseteq \R$
where $S=\{ F, R \}$. We apply the front face-rotation twice,
followed by two right-face rotations, that is
$\pi=F\*F\*R\*R$.

We just did that on a Rubik's cube app, and counted
6 instances until the initial permutation was reinstantiated.
That indicates that $\pi$ may be of order $6$
(that is $24$ quarter turns in total.)

After one $FFRR$ we may do $5$ more $\pi$ composites, that is
$FFRR FFRR FFRR FFRR FFRR,$ to revert the first move.

We could, of course, also have done $-\pi$, that is the inverse
permutation of $\pi$ (with $\pi\*-\pi=0$). $-\pi=-R\*-R\*-F\*-F,$
which with $-F=f$ and $-R=r$, may be denoted as $rrff.$

\e Now we will enumerate the three subsets,
$F\subseteq \R,$ $R\subseteq \R,$ and $F\cup R\subseteq \R.$
These subsets of permuations contribute to the abstraction of
the Rubik's cube's front face $F$ and right face $R.$  The
union of these sets $S=R\cup F$ aims to help the enumertion
of permutations which we may call (not all that seriously)
jumbles, ruffles, permutes, and sometimes but not always
rotations or composites.  These subsets together with their
operation $\*$ and their neutral element $0$ (maybe called $()$)
form subroups of the Rubik group $\R.$

\e The front face is represented by set $F,$ and set $R$
is the right face of the cube. (There is already some
ambiguity creeping and ready to interfere.
There are nouns like set, permutation, arrangement, and face;
then we see verbs such as permute, arrange, or jumble,
ruffle, or just rotate by a quarter turn; or some such
like.)

\eject
Here is the front face
$F=\{$ FUL FU FUR FR FDR FD FDL FL FC $\},$
and the right face
$R=\{$~RUF RU RUB RB RDB RD RDF RF RC $\}.$

We also see some entanglements, namely
$\{$ ( FUR RUF ), ( FDR RDF ), ( FR RF ) $\}.$
For example, FUR is the cube of cubes on the
front face inset $F,$
while RUF is the same cube (maybe called cuby)
on the right face in set $R.$
The elements FUR and RUF are also part of the
permutation (here, the action of rotating the face)
$F,$ and $R$ respectively.

\e The rotation of the front face as a permutation cycle is
$F=($ FUL FUR FDR FDL $)\*($ FU FR FD FL~$),$
and the right face re-arrangement by a clockwise quarter-turn
is $R=($ RUF RUB RDB RDF $)\*($ RU RB RD RF $),$
while the entaglements of the front and the right are
already lurking and ready to intervene.
}% end of example 2F 2R cycle
==================================================
(work said ot be in progress)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>