Rubik: a: entanglement of the faces
front, right, (FR RF), (FUR RFU), (FRD RDF)
==============================================
\beginsection 10. The example 2F 2R cycle and its abstraction

\def\R{\hbox{\bf R}}%
\e Spinning the cube.  Let $\pi\in S$ and $S\subseteq \R$
where $S=\{ F, R \}$. We apply the front face-rotation twice,
followed by two right-face rotations, that is
$\pi=F\*F\*R\*R$.

We just did that on a Rubik's cube app, and counted
6 instances until the initial permutation was reinstantiated.
That indicates that $\pi$ may be of order $6$
(that is $24$ quarter turns in total.)

After one $FFRR$ we may do $5$ more $\pi$ composites, that is
$FFRR FFRR FFRR FFRR FFRR,$ to revert the first move.

We could, of course, also have done $-\pi$, that is the inverse
permutation of $\pi$ (with $\pi\*-\pi=0$). $-\pi=-R\*-R\*-F\*-F,$
which with $-F=f$ and $-R=r$, may be denoted as $rrff.$

\e Now we will enumerate the three subsets,
$F\subseteq \R,$ $R\subseteq \R,$ and $F\cup R\subseteq \R.$
These subsets of permuations contribute to the abstraction of
the Rubik's cube's front face $F$ and right face $R.$  The
union of these sets $S=R\cup F$ aims to help the enumertion
of permutations which we may call (not all that seriously)
jumbles, ruffles, permutes, and sometimes but not always
rotations or composites.  These subsets together with their
operation $\*$ and their neutral element $0$ (maybe called $()$)
form subroups of the Rubik group $\R.$

\e The front face is represented by set $F,$ and set $R$
is the right face of the cube. (There is already some
ambiguity creeping and ready to interfere.
There are nouns like set, permutation, arrangement, and face;
then we see verbs such as permute, arrange, or jumble,
ruffle, or just rotate by a quarter turn; or some such
like.)

\eject
Here is the front face
$F=\{$ FUL FU FUR FR FDR FD FDL FL FC $\},$
and the right face
$R=\{$~RUF RU RUB RB RDB RD RDF RF RC $\}.$

We also see some entanglements, namely
$\{$ ( FUR RUF ), ( FDR RDF ), ( FR RF ) $\}.$
For example, FUR is the cube of cubes on the
front face in set $F,$
while RUF is the same cube (maybe called cuby)
on the right face in set $R.$
The elements FUR and RUF are also part of the
permutation (here, the action of rotating the face)
$F,$ and $R$ respectively.

\e The rotation of the front face as a permutation cycle is
$F=($ FUL FUR FDR FDL $)\*($ FU FR FD FL~$),$
and the right face re-arrangement by a clockwise quarter-turn
is $R=($ RUF RUB RDB RDF $)\*($ RU RB RD RF $),$
while the entaglements of the front and the right are
already lurking and ready to intervene.

\e Now we look at the FFRR permutation again.
This composite jumble has two front face quarter-turns followed
by two 90 degree turns of the right face and it gives the
permutation $F\*F\*R\*R,$ or in other more detailed words
$($ ( FUL FDR ) ( FUR FDL ) ( FU FD ) ( FR FL ) $\*$
    ( RUF RDB ) ( RUB RDF ) ( RU RD ) ( RF RB ) $)$
\eject
One way to verify this permutation (verb) is to express its
effect as a two-line expression, first $F(F\*F)$, like so
$$F(F\*F)=\left(\matrix{%
 \hbox{FUL FU FUR FR FDR FD FDL FL}\cr
 \hbox{FDR FD FDL FL FUL FU FUR FR}}\right),$$

\e %
Or in a more verbose way the right~face 180 degree
turn might look like this:

$$R(R\*R)=\left(\matrix{%
 \hbox{RFU RU RUB RB RBD RD RDF RF RC}\cr
 \hbox{RFU RU RUB RB RBD RD RDF RF RC}}\right)$$
{\def\*(#1){\circ\hbox{(#1)}}%
$$\*(RFU RBD)\*(RUB RDF)\*(RU RD)\*(RF RB)\*(RC)=$$}
$$\left(\matrix{%
 \hbox{RFU RU RUB RB RBD RD RDF RF RC}\cr
 \hbox{RBD RD RFD RF RFU RU RBU RB RC}}\right).$$

\e(to be corrected, still)

\eject
\e We try to visualise a particular face-to-face entanglement
which could be expressed within the context of the
respective edge-cubes of the front and the right~face.

\e {\def\a{\hskip 3cm}%
\def\b{\hskip 24pt}%
\def\c#1 {{\bf #1 }}%
\vskip 12pt%
\halign{#&#&#&#&#&#\cr
        &FU&&           &RU\cr
\a  FL&&\c FR &\b\c RF &&RB\cr
        &FD&&           &RD\cr}

\e {\def\a#1 #2 {\hbox{#1 #2}}%
In terms of the set of permutation cycles the entaglement may read
$\{(\a FR RF ),(\a RF FR )\},$ and again we note that
$(\a FR RF )=(\a RF FR ).$}

\e The corner-cubes have a similar entanglement. (re-thing, re-word)
{\def\a{\hskip 3cm}%
\def\b{\hskip 24pt}%
\def\c#1 {{\bf #1 }}%
\vskip 12pt%
\halign{#&#&#&#&#&#\cr
\a FLU&&\c FUR &\b\c RFU &&RUB\cr
     &Front&&&       Right\cr
\a FDL&&\c FRD &\b\c RDF &&RD\cr}}

}% end of example 2F 2R cycle
==============================================
(under re-view)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>