Rubik: a: clock-wise: direction
1234-> 2341

(1234) permutes to (2342) in aclock-wise
direction; right?

Yes and no! Because here, in the game of Rubik's cube,
we rotate _the clock_ instead of
_the dial of its clock_; so to speak.

So, if we look at the front face,
we will see the edge cubes in its standard perm,
(FU FR FD FL).

    FU
  FL  FR
    FD

Now, if we turn the cube (analoguous to the clock itself)
clock-wise, then _time_ goes back-wards and we might see

    FL
  FD  FU
    FR

Hmmm; time to re-consider the Woozy group and turn it
back also.  Maybe we will be wide-awake; eventually.

       woozy                   wide-awake
wide-awake  vacuous  ->  sleepy       woozy
       sleepy                  vacuous

(Now in progress; vucuous---for now---it seems.)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>