Rubik: a: clock-wise: TeX boxes side by side
TeX nicalities, vcenter, vbox, hbox
================================================
{\beginsection 11. Let's look at the example RUru

The move given in the example\d\ is
$($ RUru $)$ which is said to be a ``commutator.''
It is said to affect seven cubes of which two have
no part in the up~face. In our abstraction we denote
$\gamma\in G$ and $\gamma=(\hbox{ RUru }),$ with
$G=U\cup R$ and $G\subseteq R.$

Here, we start with the standard~permutation (noun)
$\sigma$ with the edge~cubes of the up~face singled out
with the expression
$\sigma_U^e=(\hbox{ UB UR UF UL }).$

We re-vise our view of the clockwise quarter turn
of the right face's edges, and therefore write
$R^e=( \hbox{ RU RF RD RB } ).$  We also appreciate, now,
that the right face's up-edge, RU, is the
same as the up face's right~edge, UR. In other words,
$\hbox{RU}=\hbox{UR}.$ (Sounds trivial, but we find this
important to keep in mind, for now.)

So, just do it! That is, let's do the first component of
or $\gamma$-move. Here it is, looking at two faces with
two expressions.
$R^2=R\*(\hbox{ RU RF RD RB }),$ giving
$R^2=( \hbox{ RF RU RB RD } ).$
Now, at the same time we have
$U^R=U\*(\hbox{ UR RF }),$ being aware that
$\hbox{UR}=\hbox{RU},$ and noting the effect that
this $R$-move has on the $U$-face which we now call
$U^R,$ and which is spelled out as
$( \hbox{ UB RF UF UL } ).$

\eject
The second component of $\gamma$ being the top~face
quarter turn, $U,$ and we may write
$U^{RU}=R\*U.$ The, admittedly verbose expression reads
$U_e^{RU}=U^R\*U,$ and then
$U_e^{R}\*U=(\hbox{ UB RF UF UL })\*U$ and we get
$U_e^{RU}=(\hbox{ UL UB RF UF }).$

We also see the entanglement that this re-arrangement
implies and we aim to express the effect is has on the $R$-face.
$R_e^R\*U^{RU}$ which shows the end~effect, a ruffle of
$(\hbox{ RF UB }),$ or some such like (for now unconfirmed.)
This makes the $R$-face to read
$R_e^{RU}=R^2\*U^{RU}$ and
$R_e^{RU}=(\hbox{ UB RU RB RD }),$ (still under re-view;
in particular our mis-conceptions, the ambiguities, and
the mis-representation are creeping all over; however
we will wittle things down just to pick up each one
of these crawlies and re-permute things to form a set
of cohesion; given time.)
\vfil\eject
Fast forward, pending completion, we have booted the Rubik
cube to do the computation $\gamma$ for us, and just note
it down in terms of edge~jumbles,
$$\gamma_R^e=(\hbox{ RF RB RD BU })$$
$$\gamma_U^e=(\hbox{ UR FR UF UL })$$

Given the standard permutaion $\sigma$ with
$\sigma_U^e=(\hbox{ UB UR UF UL }),$ and
$\sigma_R^e=(\hbox{ RU RB RD RF }),$
this allows us to calculate the components
in terms of permutation cycles
$$\gamma_R^e=(\hbox{RU RF})\*(\hbox{RB})\*(\hbox{RD})\*(\hbox{RF UL}),$$ and
$$\gamma_U^e=(\hbox{UB})\*(\hbox{UR FR})\*(\hbox{UF})\*(\hbox{UL}).$$
}% end of section 11
\eject
{\beginsection 12. Direction of the permutation
which cycles one face

\def\a#1{\hbox{ #1 }}% one cube
\def\c{\hbox{ \  }}% vacuousness
\def\d{$\*$ (clock-wise 1/4 face-turn)}% permute
\def\b#1 #2 #3 #4.{% b=left 4-typle, like  1 2 3 4.
         &\a  #1 &&         &\a  #4 \cr
\c\c\c \a #4 &\c&\a #2 &\d\a #3 &\c&\a #1\cr
         &\a  #3 &&         &\a  #2 \cr}%
\e\halign{#&#&#%
         &#&#&#\cr% a=prefix
\b 1 2 3 4.}% clock-wise rotation of a face

\e{\def\bb#1 {\hbox{#1}}%
\def\d {$\*\hbox{(FU FL FD FR)}=$}%
\halign{#&#&#%
         &#&#&#\cr% a=prefix
\b \bb FU \bb FR \bb FD \bb FL .}}%

\e{\def\bb#1 {\hbox{#1}}%
\def\d {$\*\hbox{(1432)}=$}%
\halign{#&#&#%
         &#&#&#\cr% a=prefix
\b \bb FU$_1$ \bb FR$_2$ \bb FD$_3$ \bb FL$_4$ .}}%

\e{\def\bb#1 {\hbox{#1}}%
\def\d {$\*\hbox{(1432)}=$}%
\halign{#&#&#%
         &#&#&#\cr% a=prefix
\b \bb woozy$_1$ \bb vacuous$_2$ %
\bb sleepy$_3$ \bb wide-awake$_4$ .}}%

\vfil\eject
\e The move $F$ again, with its effect on five faces.

\e{\def\bb#1 {\hbox{#1}}%
\def\d {$\*\hbox{(1432)}=$}%
\halign{#&#&#%
         &#&#&#\cr% a=prefix
\b \bb woozy$_1$ \bb vacuous$_2$ %
\bb sleepy$_3$ \bb wide-awake$_4$ .}}%

{\def\bb#1 {\hbox{#1}}%
\def\d {$\*F=$}%
\halign{#&#&#%
         &#&#&#\cr% a=prefix
\b \bb FU$_1$ \bb FR$_2$ %
\bb FD$_3$ \bb FL$_4$ .}}%

\vfil\eject\e Now, we look at some \TeX nicalities\g\/
in order to show these ruffles.
We want to utilize the $\backslash$halign
control sequence to layout our cube-faces.
We also want to use \TeX~definitions
to ease the writing of this \TeX-text.
We aim for some
yield of visual clues as to the Rubik group
model for these permutations. (re-word)

Here is the basic idea. We draw, using \TeX~text, the
four edge cubes: Up $U_1,$ Right $R_2,$ Down $D_3,$
and Left $L_4.$
{{$$\halign{#&#&#\cr%
 & $U_1$\cr%
$L_4$ && $R_2$\cr%
& $D_3$\cr}$$}}

Then with the help of \TeX\/ and its $\backslash$halign
control sequence, we abstract the drawing of this
particular face.
{\def\a{$$\halign{##&##&##\cr%
 & $U_1$\cr%
$L_4$ && $R_2$\cr%
& $D_3$\cr}$$}
\a}

\vfil\eject
Now we parameterize the four edge cubes that we see
on this face.
{\def\a#1 #2 #3 #4.{$$\halign{##&##&##\cr%
 & $#1_1$\cr%
$#4_4$ && $#2_2$\cr%
& $#3_3$\cr}$$}
\a U R D L.

This allows us, now, to facilitate these abstraction
and apply things to our favourite set.
\a \hbox{woozy} \hbox{vacuous} \hbox{sleepy} \hbox{wide-awake}.

And we may even show a plain and clear view of
these face rotations, in our own perception,
by using the set $\{1,2,3,4\}$ as the Group's set.
\a 1 2 3 4.

Let's see a representation of our adjoining
faces. These are the four faces that adjoin this one.
This one is face $(1,2,3,4).$  We say
that the face in front of us is face $0$ and
substitution yields $(01,02,03,04).$

\a 01 03 04 05.

Our edge cube $01$ is adjoined to face $1.$ (language)
and face one sees its edge~cube $10$ to be adjoined to
us. Yes! we are face $0.$

\a 12 13 10 15.

And here we take a break to reflect, and to replenish
our writing finger non-fossil, no-no-nothing fuel.

\vfil\eject\e Another approach to get these faces to show themselves side~by~side.

{\a 01 03 04 05.}

Now, then! we backup a few steps and just move staight ahead.

\e
{% \vcenter, \hbox and \vbox
$\vcenter{\halign{#&#&#\cr
&$B_1$\cr
$L_4$&&$R_2$\cr
&$F_3$\cr}}$

$\vcenter{\halign{#&#&#\cr
&$U_1$\cr
$L_4$&&$R_2$\cr
&$D_3$\cr}}$
\qquad
$\vcenter{\halign{#&#&#\cr
&$U_1$\cr
$F_4$&&$B_2$\cr
&$D_3$\cr}}$
\qquad
$\vcenter{\halign{#&#&#\cr
&$U_1$\cr
$R_4$&&$L_2$\cr
&$D_3$\cr}}$
\qquad
$\vcenter{\halign{#&#&#\cr
&$U_1$\cr
$B_4$&&$F_2$\cr
&$D_3$\cr}}$

$\vcenter{\halign{#&#&#\cr
&$F_1$\cr
$L_4$&&$R_2$\cr
&$B_3$\cr}}$
}% end of vcenter

\eject
{% \vcenter, \hbox and \vbox
\def\edges#1 #2 #3 #4.{$\vcenter{\halign{##&##&##\cr
&{#1}$_1$\cr
{#4}$_4$&&{#2}$_2$\cr
&{#3}$_3$\cr}}$}

\e We show the edges as number pairs where the first digit of each two-digit number
gives the cube's face visible on this cube of cubes' face.  The second such
digit gives the adjoining face of the standard permutation $\sigma,$ where the
faces are: $0$ to represent the front face,
$1$ its back, $2$ to the left, $3$ on the right, $4$ faces up, and $5$ looks down.

We break down $\sigma$ and show just its edges, with

\smallskip$\sigma_F^e=(02,05,03,04){}_F^e$, $\sigma_B^e=(12,14,13,15){}_B^e$,

$\sigma_U^e=(21,25,20,23){}_U^e$, $\sigma_D^e=(30,35,31,32){}_D^e$,

$\sigma_L^e=(42,40,43,41){}_L^e$, and $\sigma_R^e=(52,51,53,50){}_R^e$.
{\sl(this is unchecked and unfinished)}

\e\edges 21 25 20 23.\par\thinspace
\edges 02 05 03 04.
\qquad
\edges 52 51 53 50.
\qquad
\edges 12 14 13 15.
\qquad
\edges 42 40 43 41.\par\thinspace
\edges 30 35 31 32.}% end of vcenter

}% end of section 12
\bye
================================================
(in progress)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>