Rubik: a: clock-wise: viewing the rotation R
at different aspects
==================================================
\beginsection 13. The concrete move RUru

{\def\edges#1 #2 #3 #4.{$\vcenter{\halign{##&##&##\cr
&{#1}$_1$\cr
{#4}$_4$&&{#2}$_2$\cr
&{#3}$_3$\cr}}$}

We start with the representation of the
standard permutation $\sigma^e$ in the context of
edge cubelets.

\e
\edges 21 25 20 24.\par\thinspace
\edges 02 05 03 04.\qquad\edges 52 51 53 50.\qquad
\edges 12 14 13 15.\qquad\edges 42 40 43 41.\par\thinspace
\edges 30 35 31 34.

And we re-write in a somewhat narrative style (as we
perceive it so to be.)

\smallskip$\sigma_F^e=(02,05,03,04){}_F^e$,
$\sigma_B^e=(12,14,13,15){}_B^e$,

$\sigma_U^e=(21,25,20,24){}_U^e$,
$\sigma_D^e=(30,35,31,34){}_D^e$,

$\sigma_L^e=(42,40,43,41){}_L^e$,
and $\sigma_R^e=(52,51,53,50){}_R^e$.

\hfil\eject
Now we apply the move $R,$ which spells out to be
$(1432)_R^e$, if you know what this means.  We perceive
the meaning to be $R_1\to R_4$, $R_2\to R_1$,
$R_3\to R_2$, and $R_4\to R_3$, and we think this to
be a clockwise $1/4$ turn around the $x$ axis, that
is the axis from left to right. {\it(We admid that the
concept of direction such as left and right is a
woozy one; but it's not vacuous.)}

\smallskip
\edges 21 05 20 24.\par\thinspace
\edges 02 35 03 04.\qquad\edges 50 52 51 53.\qquad
\edges 12 14 13 25.\qquad\edges 42 40 43 41.\par\thinspace
\edges 30 15 31 34.

\smallskip
We note that with the move of the right facelets we
also see changes on the adjoining faces, namely
$F_2$ sees $05\to 35$, $B_4$ moves $15\to 25$,
$U_2$ substitues $25\to 05$, and finally $D_2$
transposes its edge $35\to 15$. (to verify)

\smallskip
Wittling things down, we see things scattered around us,
such as

\e$\{$( set, coset ), ( Definitionsmenge, Wertemenge )$\}$,

$\{$ injective, surjective, bijective $\}$.

\vfil\eject
We observe the effects of this rotation, $R^e$ as it
produces five permutaion cycles of order four, namely

\smallskip
$R_R^e=(50, 53, 51, 52)_R^e$

$R_F^e=(05, 35, 15, 25)_F^e$

$R_U^e=(15, 25, 05, 35)_U^e$

$R_B^e=(25, 05, 35, 15)_B^e$
==================================================
(progressin to search for the sense of direction)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>