Rubik: a: clock-wise: another article
solving problems, in general, the cube for example
================================================
\beginsection 14. The group of three slices

% https://books.google.co.uk/books?id=_n1vr0_RbXoC&lpg=PA786&pg=PA786
So called slices are sandwiched between opposing faces.
The group, so generated from the set of moves\break
$\{$~2R2l, 2F2b, 2U2d $\}$
is said to have 12 elements\gg.

Maybe we need a reminder of the group abstraction, and
maybe we should convince ourselves that this is a
group, and maybe we should account for all the elements,
the completeness, the associativities of the compositions,
the identiy element, and the existence of inverse elements.

\e First, we count the basic generating elements:
1. $FFbb$, 2. $RRll$, 3. $UUdd$.

Next, we compose with two basics:
4. $FFbb\*RRll$, 5. $FFbb\*UUdd$, 6. $RRll\*UUdd$

Last come compsitions from three basic generating elements:

7. $(FFbb\*RRll)\*UUdd$,
8. $(RRll\*FFbb)\*UUdd$,
9. $(FFbb\*UUdd)\*RRll$,

10.$(UUdd\*FFbb)\*RRll$,
11.$(RRll\*UUdd)\*FFbb$,
12.$(UUdd\*RRll)\*FFbb$.

{\it(This is the first attempt of enumerating this group
and the result is unverified.)}

\bye
================================================
(considering other views, and progressing in time)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>