Rubik: a: clock-wise: taking pictures
to illustrate, here, the 'spin' or orientation of
the centre-facelets/cubelets as they are effected by
the quantums of S+ group elements.

Also, different observation references, FUR for a
start.  S_F^+, for example is defined to move
opposing faces by 180 degree. However, we observe
from the said corner cube that slice_0 rotates while
we, the observer, stay in our place, front, top, right.

(under consideration)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>